more progress on SCRUM 83
